+++ oxe-web-community/src/component/Login.jsx
@@ -6,7 +6,7 @@ import { getRequest, postRequest } from "../utils/request.jsx";
 import { validatePassword, validateEmail } from "../utils/re.jsx";
 import Info from "./box/Info.jsx";
 import { getUrlParameter } from "../utils/url.jsx";
-import { getCookieOptions, getGlobalAppURL, getApiURL } from "../utils/env.jsx";
+import { getGlobalAppURL, getApiURL } from "../utils/env.jsx";
 import DialogHint from "./dialog/DialogHint.jsx";
 import Version from "./box/Version.jsx";
 
@@ -50,22 +50,21 @@ export default class Login extends React.Component {
 		// Get the token if the user reaches the app though a password reset URL
 
 		if (getUrlParameter("action") === "reset_password") {
-			// TODO use httponly cookies
-			this.props.cookies.set("access_token_cookie", getUrlParameter("token"), getCookieOptions());
+			this.props.cookies.set("access_token_cookie", getUrlParameter("token"), {});
 		}
 
 		// Log in the user if there is an existing cookie
 
 		if (getUrlParameter("action") !== "reset_password") {
-			if (this.props.cookies.get("access_token_cookie")) {
-				getRequest.call(this, "private/get_my_user", (data) => {
-					this.props.connect(data.email);
-				}, (response2) => {
+			getRequest.call(this, "private/get_my_user", (data) => {
+				this.props.connect(data.email);
+			}, (response2) => {
+				if (response2.status !== 401 && response2.status !== 422) {
 					nm.warning(response2.statusText);
-				}, (error) => {
-					nm.error(error.message);
-				});
-			}
+				}
+			}, (error) => {
+				nm.error(error.message);
+			});
 		}
 
 		// This function to notify if the password has been reset correctly
@@ -97,8 +96,7 @@ export default class Login extends React.Component {
 			password: this.state.password,
 		};
 
-		postRequest.call(this, "account/login", params, (response) => {
-			this.props.cookies.set("access_token_cookie", response.access_token, getCookieOptions());
+		postRequest.call(this, "account/login", params, () => {
 			this.props.connect(this.state.email);
 		}, (response) => {
 			nm.warning(response.statusText);
@@ -143,6 +141,7 @@ export default class Login extends React.Component {
 		};
 
 		postRequest.call(this, "account/reset_password", params, () => {
+			this.props.cookies.remove("access_token_cookie", {});
 			document.location.href = "/?reset_password=true";
 		}, (response) => {
 			nm.warning(response.statusText);
@@ -152,8 +151,9 @@ export default class Login extends React.Component {
 	}
 
 	backToLogin() {
-		this.props.cookies.remove("access_token_cookie", getCookieOptions());
-		window.location.replace("/");
+		this.props.cookies.remove("access_token_cookie", {});
+		this.setState({ view: "login" });
+		window.history.pushState({ path: "/login" }, "", "/login");
 	}
 
 	onKeyDown(event: React.KeyboardEvent<HTMLDivElement>) {