+++ oxe-web-community/src/component/Menu.jsx
@@ -6,7 +6,6 @@ import SideNav, {
 import "@trendmicro/react-sidenav/dist/react-sidenav.css";
 import { Link } from "react-router-dom";
 import Loading from "./box/Loading.jsx";
-import { getCookieOptions } from "../utils/env.jsx";
 
 export default class Menu extends React.Component {
 	constructor(props) {
@@ -38,8 +37,7 @@ export default class Menu extends React.Component {
 				className={"fade-in"}
 				onSelect={(selected) => {
 					if (selected === "disconnect") {
-						this.props.cookies.remove("access_token_cookie", getCookieOptions());
-						window.location.replace("/");
+						this.props.logout();
 					} else {
 						this.props.changeMenu(selected);
 					}